Restructure project for full export
