+++ package.json
@@ -8,8 +8,8 @@
     "url": "https://croct.com"
   },
   "license": "UNLICENSED",
-  "main": "build/index.js",
-  "types": "build/index.d.ts",
+  "main": "index.js",
+  "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/croct-tech/sdk-js.git"
@@ -23,8 +23,7 @@
     "lint": "eslint 'src/**/*.ts' 'test/**/*.ts'",
     "lint-fix": "eslint --fix --fix-type suggestion --fix-type problem 'src/**/*.ts' 'test/**/*.ts'",
     "test": "jest -c jest.config.js --coverage",
-    "rollup": "rollup -c rollup.config.js",
-    "rollup-min": "rollup -c rollup.config.js --environment minify"
+    "build": "tsc"
   },
   "devDependencies": {
     "@types/jest": "^24.0.15",
@@ -39,18 +38,13 @@
     "jest-websocket-mock": "^1.5.1",
     "mock-socket": "^8.1.1",
     "node-fetch": "^2.6.0",
-    "rollup": "^1.31.0",
-    "rollup-plugin-dts": "^1.2.1",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-typescript2": "^0.21.2",
     "temp-dir": "^2.0.0",
     "ts-jest": "^25.4.0",
     "typescript": "^3.7.5"
   },
   "files": [
-    "build/index.js",
-    "build/index.d.ts"
+    "*.js",
+    "*.ts"
   ],
   "browserslist": [
     "last 1 version"
