+++ src/channel/index.ts
@@ -1,4 +1,4 @@
-interface Closeable {
+export interface Closeable {
     close(): Promise<void>;
 }
