+++ src/sdk.ts
@@ -1,6 +1,6 @@
 import {Container} from './container';
 import Context, {TokenScope} from './context';
-import Logger from './logger';
+import Logger from './logging';
 import {BOOTSTRAP_ENDPOINT_URL, EVALUATION_ENDPOINT_URL, TRACKER_ENDPOINT_URL, VERSION} from './constants';
 import {configurationSchema} from './schema/sdkSchemas';
 import {formatCause} from './error';
