+++ src/token/replicatedStorage.ts
@@ -1,4 +1,4 @@
-import Token, {TokenStorage} from '../token';
+import Token, {TokenStorage} from './index';
 
 export default class ReplicatedStorage implements TokenStorage {
     private primary: TokenStorage;
