+++ test/retry/neverPolicy.test.ts
@@ -1,4 +1,4 @@
-import NeverPolicy from '../../src/retryPolicy/neverPolicy';
+import NeverPolicy from '../../src/retry/neverPolicy';
 
 describe('A never policy', () => {
     test('should impose an infinity delay', () => {
