+++ test/token/index.test.ts
@@ -1,5 +1,5 @@
-import Token, {FixedTokenProvider} from '../src/token';
-import {base64UrlEncode} from '../src/base64Url';
+import Token, {FixedTokenProvider} from '../../src/token';
+import {base64UrlEncode} from '../../src/base64Url';
 
 describe('A token', () => {
     const appId = '7e9d59a9-e4b3-45d4-b1c7-48287f1e5e8a';
