+++ test/token/replicatedStorage.test.ts
@@ -1,6 +1,6 @@
 import Token from '../../src/token';
-import ReplicatedStorage from '../../src/tokenStorage/replicatedStorage';
-import InMemoryStorage from '../../src/tokenStorage/inMemoryStorage';
+import ReplicatedStorage from '../../src/token/replicatedStorage';
+import InMemoryStorage from '../../src/token/inMemoryStorage';
 
 describe('A replicated storage', () => {
     const token = Token.issue('7e9d59a9-e4b3-45d4-b1c7-48287f1e5e8a', 'c4r0l', 1440982923);
