+++ tsconfig.json
@@ -17,7 +17,6 @@
     "outDir": "build",
     "downlevelIteration": true,
     "declaration": true,
-    "declarationDir": "build/declarations",
     "lib": ["es6", "dom", "dom.iterable", "es2018.promise"]
   },
   "include": [
